requirements.txt
streamlit==1.32.0
boto3==1.34.69
python-dotenv==1.0.1 